Move tabbed menu concepts into a base class
